+++ src/components/layout/Backdrop/index.tsx
@@ -5,19 +5,15 @@ import ReactDOM from 'react-dom'
 
 const useStyles = makeStyles({
   root: {
-    zIndex: 1300,
+    zIndex: 0,
     top: '52px',
   },
 })
 
-const BackdropLayout = (props: { isOpen: boolean }): ReactElement | null => {
+const BackdropLayout = ({ isOpen }: { isOpen: boolean }): ReactElement | null => {
   const classes = useStyles()
 
-  if (!props.isOpen) {
-    return null
-  }
-
-  return ReactDOM.createPortal(<Backdrop classes={{ root: classes.root }} open />, document.body)
+  return ReactDOM.createPortal(<Backdrop classes={{ root: classes.root }} open={isOpen} />, document.body)
 }
 
 export default BackdropLayout
